+++ examples/src/bin/spi_loopback_dma_psram.rs
@@ -25,7 +25,7 @@
 use esp_backtrace as _;
 use esp_hal::{
     delay::Delay,
-    dma::{Dma, DmaBufBlkSize, DmaRxBuf, DmaTxBuf},
+    dma::{DmaBufBlkSize, DmaRxBuf, DmaTxBuf},
     peripheral::Peripheral,
     prelude::*,
     spi::{
@@ -67,9 +67,6 @@ fn main() -> ! {
     let miso = unsafe { mosi.clone_unchecked() };
     let cs = peripherals.GPIO38;
 
-    let dma = Dma::new(peripherals.DMA);
-    let dma_channel = dma.channel0;
-
     let (_, tx_descriptors) =
         esp_hal::dma_descriptors_chunk_size!(0, DMA_BUFFER_SIZE, DMA_CHUNK_SIZE);
     let tx_buffer = dma_alloc_buffer!(DMA_BUFFER_SIZE, DMA_ALIGNMENT as usize);
@@ -103,7 +100,7 @@ fn main() -> ! {
     .with_miso(miso)
     .with_mosi(mosi)
     .with_cs(cs)
-    .with_dma(dma_channel);
+    .with_dma(peripherals.DMA_CH0);
 
     delay.delay_millis(100); // delay to let the above messages display